requirements.txt
google-genai
langgraph
langchain-core
requests
pylint
pytest
pytest-cov
smtplib-ssl
python-dotenv
